add mne based bem script for forward solutions
In order to complete the present anatomical pipeline for complete
simulated modalities in TVB, forward solutions for MEG and EEG
require that the head model be formed and solved. This script
uses the FreeSurfer segmentation and MNE command line tools to
build the BEM and solve it. It is still necessary to manually
align sensors to MRI space, but after this step, forward solutions
are trivial to obtain using the mne-python library.